Moved interface declarations to new types.ts file
