+++ frontend/src/Layout/RootLayout.tsx
@@ -2,12 +2,9 @@ import { Outlet } from 'react-router-dom';
 import Nav from './Components/Nav/Nav';
 import Footer from './Components/Footer';
 import Box from '@mui/material/Box';
-interface RootLayoutProps {
-    isLoggedIn: boolean;
-    setIsLoggedIn: (value: boolean) => void;
-}
+import { RootLayoutProps } from '../types/types';
 
-const RootLayout: React.FC<RootLayoutProps> = ({ isLoggedIn, setIsLoggedIn }) => {
+const RootLayout = ({ isLoggedIn, setIsLoggedIn }: RootLayoutProps) => {
 
     return (
         <Box sx={{ display: 'flex', flexDirection: 'column', justifyContent: 'space-between', minHeight: '100vh', minWidth: '320px' }}>
